[4] Add existing source code
The source code from the MolDyn_NVE.tar.gz archive was adapted into the class
molecular_dynamics_sim. The result is a code that it is far from ready,
as it still needs a lot of adjustments and corrections in order to work
properly.
Since we have now more than three object files, the makefile was also
rewritten using automatic variables in the patterns: this will make it
also work with the future exercises without making any change.