[4] Rewrite initialisation methods
The methods in the 'molecular_dynamics_sim' class that gather the input
parameters and initialise the configuration were rewritten, using STL
libraries and possibly generalising the code to an arbitrary number of
dimensions.
The methods 'move' and 'force' were tidied up, too, in order to complete
the initialisation procedure.
The bit regarding the sampling from a Maxwell-Boltzmann distribution is
still incomplete; it will require more time.