[4] Rewrite measure() method
The method which calculates some physical quantities of the actual state
of the system was rewritten to use STL libraries and more meaningful
variable names.
Since this method does already the job of writing the "measurements"
several files, the various methods that simply returned the single
measurement are not needed anymore.