[4] Redesign input methods
The constructors and the input methods such as initialise_uniform() were
reorganised in order to avoid code repetition and to write clearer and
cleaner functions.

The parameter input process was delegated to a series of set_...()
methods, rather than the constructor: the class should not dictate how
to input the initial parameters; it should just provide an interface to
initialise them; the particular implementation of the input process is
to be left to the program using the class.

The files with the initial positions, however, are crucial to the
simulation and there is essentially only one way to input it, that is
via properly formatted text files.
Therefore it is appropriate to design a constructor which creates the
list of the particles positions from this kind of file.

The algorithm to rescale the velocities is something that was shared
between methods of the class, e.g. initialise_uniform() and the two-file
constructor; this is an unnecessary repetition, thus the algorithm was
moved to a method of its own. That method can be called as needed after
the velocity distribution has been generated.