[4] Generalise write_config_xyz() method
This method forced the user to use a fixed pattern for the output file
name should they choose the XYZ format.
This choice should be left to the user, not to the class, so as it is
more flexible in its usage, therefore the method now uses a generic
filename.
The main function in exercise04.cc was changed accordingly.