[4] Fix wrong use of integration step variable
The class methods used, in the definition of the various functions of
the Verlet algorithm, the variable time_step representing the discrete
increase in time at each new step.
This variable, however, was defined in the header but never initialised,
and it defaulted to zero: as a consequence the particle never moved
while the algorithm advanced.
The correct variable to use is integration_step, which is properly
initialised by the method set_integration_step() to be called in the
main function.