[4] Fix wrong call to std::inner_product()
The std::inner_product calls which computed e.g. the square of the
velocities in the computation of the temperature were made (erroneously)
switching the end and begin iterators in the second and third arguments.
Consequently, the function calculated the inner product of an empty
range: this explains why the distances in the potential energy
computation were always zero, for example.