[4] Break down measure() into smaller methods
The measure() method did too many things at once, calculating four physical
quantities simultaneously and writing them into external files, of which
the user could not even choose the name.
For this reason the method was broken down into three smaller methods,
each of which returns one quantity only (there is no need to define a
method which calculates the total energy: the user can just sum the
potential and the kinetic ones together).
The output procedure is left entirely to the end user.
This gives the user more flexibility on how to handle the "measurement
process".

Since both the temperature and the kinetic energy require the
calculation of the mean square velocity of the system, the new methods
were designed so as not to repeat the calculation if they are called
without advancing the algorithm to the next integration step: this
required the use of the boolean flag ms_velocity_already_computed.