[4] Implement initialise_maxwellboltzmann() method
As an alternative to sampling the initial velocities uniformly and then
rescaling them to match a desired temperature, this method samples them
directly from a Maxwell-Boltzmann distribution at the given temperature,
so there's no need to rescale them afterwards.
This sampling, however, still requires a translation of the velocities
so that the centre of mass of the system is at rest.

The main function in exercise04.cc was changed to use this
initialisation instead of the uniform one.