[5] Add Metropolis algorithm code
The code for the exercise 5.1 contains the definitions of the base
class metropolis and its derived metropolis_uniform, which implement the
Metropolis algorithm, and the code to sample the probability density of
the position of the hydrogen atom ground state.
The metropolis classes are templated on the number of dimensions, this
way the algorithm can be used for functions of an arbitrary number of
(real) variables.

The algorithm can be improved a lot, for example introducing a tool to
discard the first samples when the process has yet to equilibrate
itself; for now, this is really a basic version.