[4] Add command-line argument parsing
The program was modified to accept two command-line arguments, which are
the type of the particle and the phase of the system.
This way the program can be executed many times with different
parameters without recompiling it, and the input/output files can be neatly
organised in subfolders (named accordingly to the arguments passed to the
program).