[6] Add basic ising_1d_sim class
Parts of the code in Monte_Carlo_ISING_1D.cc file were refactored and
moved to the main function in exercise06.cc, particularly those
regarding the input procedure and the measurement of the physical
quantities.
The tasks of creation and subsequent handling of the Ising model (i.e.
of the vector of integers representing the spins) were abstracted in a
new class, ising_1d_sim, in order to avoid duplicating code every time
we need to follow the evolution of a model.

The algorithms to evolve the system to a new state, either with a
standard Metropolis or a Gibbs algorithm, were implemented as methods of
this class.

As of this commit, the code compiles successfully, but it is not known
if it really does its job.