[6] Add code to calculate physical quantities
The program was expanded in order to calculate all four requested
physical quantites of the model.
Since they cannot be directly calculated during the simulation (they
need the average values of the energy or the sum of all spins) a
different approach was needed to calculate their uncertainties; this
called for two separate functions block_averages and
block_uncertainties, instead of the usual block_statistics, to be used
when dividing the raw data from the simulation in blocks, and deriving
the average values and their errors.