[6] Fix transition probability in Gibbs algorithm
The probability that determined whether the spin was to be -1 or 1 in
the new state was wrong: the code was using the probability for the
positive spin to simulate the "spin is -1" event, that is, the
probabilities were exchanged (a minus sign was missing).
This gave as a result that the average energy had the wrong sign, since
this practically this sign error was equivalent to saying that the
energies of the 1 and -1 states were exchanged: the system, therefore,
evolved towards a positive energy state thinking it was a negative
energy state.