[6] Simplify Metropolis acceptance condition
In order to check if the acceptance threshold is less than 1 it is not
necessary to calculate it: it suffices to know whether the energy
difference between the next and previous state is positive.
By testing this instead of the acceptance threshold, the program can
skip the calculation of the latter when it is not needed.