[7] Add basic program and metropolis_NVT class
The class methods are written in a very "direct" way and there is really
a lot to improve, but at least the program calculates (correctly, it
seems) the potential energy and the pressure of the system.

A lot of routines in the code are derived from the analogues of
exercise 4.