[7] Generalise metropolis_NVT to any dimension
The implementation of the metropolis_NVT class was reworked in order to
generalise it to any dimension, so as to be able to use it to analyse a
broader class of systems.
Iterators, ranges and STL algorithms were used instead of basic C-style
loops in order to avoid explicitly mentioning anywhere in the code the
number of coordinates of the points.
Operators on vectors were defined to make the syntax simpler.