[7] Add initialisation list in metropolis_NVT ctor
The class constructor was rewritten in order to initialise the most
possible members at the beginning, instead of setting them equal to
something later.
The list in the class header in which the private members were declared
was reordered to match the initialisation list.
Some parameters that were not used throughout the class methods were
deleted.