[7] Fix transition probability
The transition probability used was the opposite of the correct one:
the system, therefore, evolved towards the "wrong" state at each step,
and the physical quantities such as the potential energy diverged.
This correction required to fix the variance of the step given in the
input files, since the correct value to reach an acceptance rate of
nearly 1/2 was very different if the energy difference between the old
and the proposed state of the system isn't the correct one.
The plots in the notebook were regenerated, this time using 1e5 steps.