[7] Add measure() methods in metropolis_NVT class
Calculating the potential energy, the pressure and the radial
distribution requires a loop over all pairs of particles, therefore
using separate methods is really inefficient if we need to calculate
more than one of them at each step.
These new measure() methods compute all quantities of interest in a
single loop, eliminating all the useless repetitions, at the cost of
some code duplication.

Now that the program computes (more) efficiently the three quantities,
the main function was modified in order to output all of them on files.
Where it made sense, headers were added in the first row of these output
files, in order to tell the reader what the columns represent; the
notebook was updated to make the input functions ignore this row.