[8] Add exercise 08.2
The program from a previous exercise was copied and adapted to solve
exercise 08.2.
The results were analysed in the Python scripts in the notebook, but
there is still a lot to do to improve them.
For example, the repeated execution of the program is very slow, and
maybe the repeated calculations with different values of mu and sigma
can be done in a single run of the program, reducing the total execution
time.