[9] Fix replacement of old generation with new one
The algorithm, after creating the two new chromosomes, just replaced
their parents with them and "completed" the new generation like that.
Instead, it should create an entirely new configuration, replacing all
old chromosomes each time a new generation is complete.
The algorithm, therefore, now fills up a vector of new chromosomes until
it reaches the appropriate size of a configuration, then it replaces the
old configuration with this vector.

In order to do this, the child chromosomes can't simply replace their
parents after they are generated in the crossover procedure; the
mutations must operate on the new chromosomes, not directly on the old
ones.