[9] Print best fit of all generations
The main function was modified so that the fittest chromosome is printed
(along with the generation number) every time a new generation is
created. This allows other programs to analyse the whole evolution of
the genetic algorithm; if they need the result only, they may just read
the last line of the output file.