[9] Include <numeric> and <functional> headers
The code compiled fine on the lab PCs, but not on my laptop, since some
functions (like std::iota) are found in the <numeric> and <functional>
headers. It is likely that the lab PCs use an old g++ compiler version
whose old libraries are different from mine. The program should now
compile on both computers, and hopefully on all computers.

My laptop also warns me of misleading indendation in an if-else block,
which is no problem at all, but I corrected anyway in order not to see
that warning.