[9] Add notebook with first results
Now that the algorithm seems to work as it should, the code was modified
to make it print, at each new generation, the length of the best path
(represented by the fittest chromosome) and of the average length of the
fittest chromosome (the best half of the set).
A notebook was added, showing the results.