[8] Add exercise08-2-grid
Exercise 08.2 requires finding the minimum of the expectation value of
the Hamiltonian, which I chose to do by plotting a grid of values for
different values of mu and sigma; this is a quick task for the program,
but hundreds of calls to ./exercise08.2 are needed to draw a nice plot,
and python is really slow when calling the executable.
I moved the repetition of the algorithm completely internal to che C++
executable, so that it is called only once, thereby greatly reducing the
overhead of the python script.
The notebook was modified accordingly to work with the new files.