Re-design random class
The previously called Random class (the random number generator provided
by the teacher) was re-designed.

The initialisation routine that appeared at the beginning of every main
function of the exercises code very much resembled what someone would
write in a constructor, so it was moved in the class constructor.

For the same reason, the "save seed" routine always appeared at the end
of each program, and it seems to me it acts as a destructor, since it
executes some code when the rng has ended its use; therefore I moved it
into the destructor.
As a bonus, this eliminates the need to explicitly call the function at
the end of the program, something that I frequently forgot to do in
later exercises.

The syntax was adapted to the style of the code written by myself, i.e.
using snake_case and avoiding capitalisation of class and method names,
starting a "{" block on a new line, and so on.

The class was put into a new namespace, "lsn", since there already
exists a function called "random" in the base language.

Some method names were changed. I don't know what "Rannyu" stood for,
but I think "uniform" is a way more informative name for a method that
returns a uniformly distributed number.