Rewrite block_statistics function using iterators
The function had a std::vector<double> as the variable holding the data
to be processed. This made the function code easy to read, but it was
obviously very limitating since there was really no other choice on how
to store the input data, e.g. one could not use an array.
The new code accepts a generic iterator over a container, both for input
and output data, allowing the function to be used (at the very least)
with any STL container.
The code is more complicated (but not so much, after all) but this is an
acceptable price to pay to make the function work in a more general
context.