Move statistics.hh in common "include" directory
The header (just like random.hh, which will follow soon the same fate)
is meant to be used by the programs in every exercise, so in order to
avoid copying the file in every directory each time it is modified it
has been placed in a common directory at the root of the project.
The code for exercise 1.1, on which the new function was tested,
was edited to reflect this new directory structure.