[11] Add basic neural network for exercise 11.3
A basic solution to exercise 11.3, consisting of a simple neural network
as in the previous exercises, was written in the notebook.
I have yet to find a decent way to superimpose the scatter plot of the
predicted data to the target function, for now I'll have to leave it as
it is, even if it is really hard to read the plot.