Move "random" class definition to shared folder
Instead of each exercise containing a copy of the header and
implementation files, the code structure was refactored in order to use
a single .hh and .cc file, stored in a shared folder.
This makes propagating the changes to every program, if one should
change the definition of the class in some way, totally effortless.