Use exceptions instead of asserts in "random" code
In the implementation of the "random" class the asserts that made sure
that some parameters were positive were replaced by exceptions.
This should make the error messages more expressive. Moreover, since
I don't believe that using a negative parameters in such cases is
strictly an implementation error, but is something that is entirely
up to the users of the class, an exception should be the more logical
way to handle the error.