[1] Complete exercise 1
The notebook was redesigned in order to make the plots prettier, but
with fewer lines of code.
I removed the 'TeX' option for matplotlib, since I don't need the full
TeX engine to draw the plots, the basic layer is enough to typeset the
formulae in the titles and labels.
Some parameters in the programs were adjusted, and the output files of
each one were added to the repository.
Unless I find some errors later, this exercise should be now complete.