[6] Rewrite statistics functions with iterators
The two functions that calculate the average and uncertainty using the
usual blocking technique were rewritten using templates and iterator,
just as I wrote the block_statistics function for the other exercises.
This led to some other changes in the code, such as the introduction of
the block_size variable: for this reason the input file formal was
overhauled.