[6] Fix measured quantities vectors mix-up
While getting the measured quantities from the simulator classes,
somehow the sum of all spins got saved in the vector representing the
square of the spin sum. A vector holding this sum when the magnetic
field is zero was also missing.
This was fixed and now the results are okay.