[6] Add remaining plots to exercise notebook
The plots obtained with the Metropolis sampling technique were added to
the notebook, that was also redesigned in order to group all plots in a
single cell.
The code was a bit rewritten so as to make it clearer and cleaner.
The output files of the program, both with the Metropolis and Gibbs
sampling techniques, were added to the index.