[6] Shuffle order of spins to be flipped
When the spins are selected, in the order as they appear in the list,
to be flipped to generate a new state, the results of the simulation
are really bad, just unacceptable. It is likely that this way of
selecting the spins introduces some important correlations in the
process.
I changed this by shuffling the selection order of the spins, and
looking at the results it seems that this was the correct intuition.
Now the results are similar to the ones obtained with Gibbs sampling.