[6] Complete exercise 6
With the program fully functional and the bugs fixed, the remaining
corrected plots were added to the notebook, and the parameters of the
simulations were updated to match the values I used.
I wrote a small comment at the beginning of the notebook to explain
briefly what I did in this exercise.
The original source code given by the professor is not needed
anymore so I deleted the files.
The exercise should be now complete.