[7] Complete exercise 7.1
The notebook was completed with the addition of some comments to the
results, and redesigning the structure to be clearer.
I added a small help message to the programs in case that the input
command line arguments are wrong.
I added the input and output files to the index.
This part of the exercise should be complete, now.