[7] Redesigned input formats
The input format of the input.dat files (that were added to the index)
was changed in order to specify the block size, so that I can use the
optimal block size calculated from the analysis of the autocorrelation
function I did.
I added a help message to indicate how command line arguments should be
passed to the program.
Since the program uses the Verlet algorithm, it was renamed with a
_verlet suffix, thus it can be distinguished from the program using
the Metropolis sampling that I will add later.