[7] Join exercise 7.2, 7.3 and 7.4
Exercises 7.2 and 7.3 require no plots to be shown, just the addition of
some functions to the code. Since these function are present in the
programs of exercise 7.4 anyway, there is no need to have a separate
directory for exercises 7.2 and 7.3.
For this reason, those two directories were deleted, and their files
moved together with the exercise 7.4 code.

The input.dat files were edited so that they could be used with both
_metropolis and _verlet programs (_metropolis needs one line more that
tells the equilibration steps, that can be ignored by _verlet).