[7] Redesigned input formats (again)
The simulation parameters do not depend on which element the particles
are made of: both argon and krypton are described by the very same
simulation. Their differences are accounted for in the notebook, when
the simulated data need to be rescaled in SI units, e.g. the unit of
energy is different between the two elements.
For this reason, the separated "argon" and "krypton" directories are
replaced by the unified directories that distinguish only the phases of
the system.
In those directories, the input files were redesigned to account for the
different parameters the Verlet and Metropolis algorithms need: it would
be too difficult to use a single input file for both, so I created a
file for each case.
This also accidentally solves the problem that the Metropolis algorithm
interpreted the integration step size as the standard deviation used
when generating the proposed step, thereby producing a too high
acceptance ratio and nonsensical results.