[7] Redesigned notebook
The data produced by the programs in exercise 7.2 and 7.3 was removed,
since it is not requested by the exercise.
What remains was better organized so that it produced, in fewer space
and with cleaner code, more informative plots, e.g. I grouped those
referring to the same quantity generated by the Metropolis and Verlet
algorithms, so that they can be easily compared.
I also added instructions on how to recreate the results.