[8] Complete exercise 8
After having removed from the notebook the parts that were not requested
by the exercise, the rest was better organised by naming the executables
and the output files as _flat or _bumps according to the used trial
wavefunction.
In general there is a copy of the executables for each of the two cases,
so that I don't have to manually change the code every time I want to
switch the wavefunction.

The input and output files were added to the index, and now the exercise
should be complete.