[10] Complete exercise 10
The output files were organised according to the type ("circle" or
"square") of the problem they refer to, to keep the directory clean.
The notebook was polished, all necessary files were added to the index
and the makefile was corrected so that it compiles the code using mpic++
instead of g++ alone.
With these changes, the exercise should now be complete.