[11] Complete exercise 11
With some final tweaks to the parameters (particularly the noise level)
in exercise 11.3, I was able to comment the performance of the trained
networks in more depth. Unfortunately it was hard to track down which
one performed best in all cases since there are too many parameters one
could change, but at least I was able to point out some general features
that affect the results.
With this last change, and some minor aesthetical changes in the other
notebooks, I believe the exercise is now complete.