[4] Remove dependencies on argon or krypton
Since the description of the "Lennard-Jones system" doesn't depend,
when expressed in reduced units, on the particular element the particles
are made of, the separated simulations for argon and krypton were
"united" in one simulation per phase only.
The notebook and the executable files were changed to reflect this,
copying from the files of exercise 7.4 (that implemented the same
algorithm).