[4] Add molecular_dynamics_sim from ex. 7 methods
In order to simplify the code, some methods (such as the arithmetic
operations on vectors, or the quotient function using a vector as
parameter) were copied from the implementation of the
molecular_dynamics_sim class in exercise 7.4.
Some loops were also rewritten using iterators in place of plain
indices, making use of these new methods.