[4,7] Fix divide-by-zero bug in class ctor
In the programs implementing the Verlet algorithm, for some reason,
after the second "dynamo" class loaded the two configurations I
re-initialised the velocities of the particles. This rendered the
initialisation with two config files practically useless, so I removed
the second re-initialisation.
This uncovered a serious bug in the ctor with two config files, that
ran the move() method, but at that time the integration_step variable
was not yet initialised, so it was zero, the initial velocities were all
infinity and this produced a cascade of problems in the subsequent
calculations (there were nan's everywhere).
That second re-initialisation hid this problem since it generated the
velocities anew, regardless of the integration_steps.

In this commit I also fixed the sampling of the velocities in the
initialise_boltzmann method: the function used a normal distribution
with the temperature T as a standard deviation, while really it should
have been sqrt(T).