[4,7] Fix force calculation in move() algorithm
In the move() method in the molecular_dynamics_sim class, the force was
calculated during the computation of the new positions of each particle
(i.e. in the very same loop).
This caused wrong results since the positions were updated when the
calculation of the forces was not yet complete.
Since the forces are supposed to be calculated using the current
positions of the particle, this had the consequence that the second
particle measured a force with the first particle in its new position,
and so on. The last particle in the list saw all the other particles
already in their new positions.
This may have caused some divergences in the physical quantities,
because the updated positions may overlap with the previous ones.

This was fixed by moving the calculation of the forces in a separate
loop, that executes before any computation of the new positions begins.