[4,7] Simplify calculation of forces
Since the computation of the square distance is the same for all
components of the force vector, the only thing that differs being the
component of the displacement vector that multiplies the output, it can
be reused if I calculate all three components at once and return them as
a vector.
This should speed up the simulation a little.