[5] Add metropolis_normal sampler class
As requested by the exercise, I should study how the convergence of the
average radius changes if I use a normal (instead of uniform)
distribution to sample the new points, therefore I implemented a new
class derived from metropolis that uses a Gaussian kernel.
The main program was changed to make use of this new class, and the
output files were a bit simplified.
The notebook was updated with the new results.