[5] Add different starting points
Using an input file, the user can now choose whether to start from a
point close to the origin of far from it.
These are two predefined options but the user can define any other as
long as they respect the syntax of the program (which maybe I'll have to
define clearly in the notebook).
The notebook was updated with the results of the two cases.