[5] Implement equilibration of samples
The program was modified to read, from the input file, two parameters
that indicate the number of equilibration steps for the 1s or 2p
algorithm.
The actual values used in the two predefined simulations were obtained
from the plots in the first part of the notebook.

The notebook was updated using the results of the equilibrated samples:
I reorganised the cells a bit in order to group the plots of the same
quantities, always drawing one for the system starting closer to the
origin, and one for the other one.