[5] Re-compute average radius
With the information about the equilibration time and the
autocorrelation length obtained in the previous steps, the average
radius was re-calculated using the optimal parameters.
The plots in the last part of the notebook were updated with the new
data, using now 1e6 steps to simulate the system as requested by the
exercise. Since the file containing the sampled positions would be too
big (approximately 40 MB each) to fit in the home directory of the
lab account, I decided to record the position once every 20 steps.